feat: integrations tests implemented

+++ src/controllers/mealController.ts
@@ -1,12 +1,14 @@
 import { NextFunction, Request, Response } from "express";
-import mealRepository from "../repositories/mealRepository.js";
-import { CreateMealParams } from "../protocols/mealProtocol.js";
+import mealRepository from "../repositories/mealRepository";
+import { CreateMealParams } from "../protocols/mealProtocol";
+import httpStatus from "http-status";
+import { badRequestError } from "../errors/index";
 
 async function create(req: Request, res: Response, next: NextFunction) {
   const newMeal: CreateMealParams = req.body;
   try {
     await mealRepository.createMeal(newMeal);
-    res.sendStatus(201);
+    res.sendStatus(httpStatus.CREATED);
   } catch (err) {
     return next(err);
   }
@@ -15,7 +17,7 @@ async function create(req: Request, res: Response, next: NextFunction) {
 async function getAll(req: Request, res: Response, next: NextFunction) {
   try {
     const meals = await mealRepository.getAll();
-    res.status(200).send(meals);
+    res.status(httpStatus.OK).send(meals);
   } catch (err) {
     return next(err);
   }
@@ -24,9 +26,10 @@ async function getAll(req: Request, res: Response, next: NextFunction) {
 async function update(req: Request, res: Response, next: NextFunction) {
   const id: number = parseInt(req.params.id);
   try {
+    if (isNaN(parseInt(req.params.id))) throw badRequestError();
     await mealRepository.updateMeal(id);
     const meals = await mealRepository.getAll();
-    res.status(200).send(meals);
+    res.status(httpStatus.OK).send(meals);
   } catch (err) {
     return next(err);
   }
@@ -35,9 +38,10 @@ async function update(req: Request, res: Response, next: NextFunction) {
 async function remove(req: Request, res: Response, next: NextFunction) {
   const id: number = parseInt(req.params.id);
   try {
+    if (isNaN(parseInt(req.params.id))) throw badRequestError();
     await mealRepository.removeMeal(id);
     const meals = await mealRepository.getAll();
-    res.status(200).send(meals);
+    res.status(httpStatus.OK).send(meals);
   } catch (err) {
     return next(err);
   }
